--- src/test/java/com/qwazr/jdbc/cache/OnDiskCacheDerbyTest.java
@@ -1,58 +1,0 @@
-/**
- * Copyright 2016 Emmanuel Keller / QWAZR
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.qwazr.jdbc.cache;
-
-import org.junit.FixMethodOrder;
-import org.junit.runners.MethodSorters;
-
-import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.sql.ResultSet;
-
-import static org.junit.Assert.fail;
-
-@FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class OnDiskCacheDerbyTest extends DerbyTest {
-    private static String jdbcCacheUrl = null;
-
-    @Override
-    Class<? extends ResultSet> expectedResultSetClass() {
-        return CachedOnDiskResultSet.class;
-    }
-
-    @Override
-    String getOrSetJdbcCacheUrl() {
-        if (jdbcCacheUrl == null) {
-            String tempDirPath;
-            try {
-                tempDirPath = Files.createTempDirectory("jdbc-cache-test").toUri().getPath();
-                if (tempDirPath.contains(":") && tempDirPath.startsWith("/"))
-                    tempDirPath = tempDirPath.substring(1);
-                jdbcCacheUrl = "jdbc:cache:file:" + tempDirPath + File.separatorChar + "cache";
-            } catch (IOException e) {
-                fail("Can not create the cache dir: " + e.getMessage());
-                return null;
-            }
-        }
-        return jdbcCacheUrl;
-    }
-
-    @Override
-    String getDerbyDbName() {
-        return "myDB1";
-    }
-}
