+++ src/test/java/com/qwazr/jdbc/cache/OnDiskCacheJdbcWithDerbyBackendTest.java
@@ -1,0 +1,39 @@
+/**
+ * Copyright 2016 Emmanuel Keller / QWAZR
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.qwazr.jdbc.cache;
+
+import org.junit.BeforeClass;
+
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+
+public abstract class OnDiskCacheJdbcWithDerbyBackendTest extends JdbcWithDerbyBackendTest {
+    private static String tempDirPath;
+
+    @BeforeClass
+    public static void createTmpDir() throws IOException {
+        tempDirPath = Files.createTempDirectory("jdbc-cache-test").toUri().getPath();
+        if (tempDirPath.contains(":") && tempDirPath.startsWith("/")) {
+            tempDirPath = tempDirPath.substring(1);
+        }
+    }
+
+    @Override
+    String getOrSetJdbcCacheUrl() {
+        return "jdbc:cache:file:" + tempDirPath + File.separatorChar + "cache";
+    }
+}
